requirements.txt
absl-py==0.8.1
alembic==1.3.1
anaconda-client==1.7.2
astor==0.7.1
astroid==2.0.4
backcall==0.1.0
bleach==2.1.4
certifi==2018.8.24
cffi==1.13.2
Click==7.0
cloudpickle==0.8.1
clyent==1.2.2
colorama==0.3.9
cycler==0.10.0
dask==1.2.0
decorator==4.3.0
entrypoints==0.2.3
enum34==1.1.6
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-MySQL==1.4.0
Flask-SQLAlchemy==2.4.1
gast==0.3.3
gevent==1.4.0
greenlet==0.4.15
grpcio==1.12.1
h5py==2.8.0
html5lib==1.0.1
imageio==2.4.1
ipykernel==4.10.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
isort==4.3.4
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10.3
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
jupyter-http-over-ws==0.0.6
# Editable install with no version control (Keras==2.2.3)
-e c:\users\gunaya\anaconda3\envs\deeplearning\lib\site-packages
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
livereload==2.6.0
Mako==1.1.0
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==3.0.0
mccabe==0.6.1
mistune==0.8.3
mkdocs==1.0.4
mkl-fft==1.0.6
mkl-random==1.0.1
more-itertools==6.0.0
mysql-connector==2.2.9
nb-anacondacloud==1.4.0
nb-conda==2.2.1
nb-conda-kernels==2.1.0
nbconvert==5.3.1
nbformat==4.4.0
nbpresent==3.0.2
networkx==2.3
notebook==5.5.0
numpy==1.17.4
olefile==0.46
opencv-python==4.0.0.21
opt-einsum==3.1.0
pandocfilters==1.4.2
parso==0.3.1
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
protobuf==3.10.0
pycparser==2.19
Pygments==2.2.0
pygpu==0.7.6
pylint==2.1.1
PyMySQL==0.9.3
pyparsing==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2018.9
PyWavelets==1.0.0
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
requests==2.14.2
scikit-image==0.14.0
scikit-learn==0.20.0
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.3.11
tb-nightly==1.14.0a20190603
TBB==0.1
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
Theano==1.0.2+2.gc449c8699
toolz==0.9.0
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.4.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
